Move Logger type down to an Internal module
This will allow some modules that want to depend on it, but that Logger
depends on, make use of it -- such as to add Logger-specific behavior to
colors.
